hvcC: patch the produced init segment, not just the input extradata
The end-to-end test added here reads the hvcC out of the served init.mp4
and asserts there is nothing left to normalize. It failed: FFmpeg's mp4
muxer does not copy our extradata into the sample entry — ff_isom_write_hvcc
rebuilds the record from the parameter sets it collected and writes
array_completeness = 0, while movenc still names the entry hvc1. So the
input-side rewrite never reached the artifact AVPlayer actually parses.

HVCCNormalizer.patch(initSegment:) walks the box tree to the hvcC payload
and normalizes it there, and both init-write sites in the remuxer now go
through one helper that applies it. The splice is refused outright if the
normalized payload changed length, since that would move every following
box and invalidate the enclosing sizes — the arrays that would shrink it
are already gone by then, dropped on the input side.

Keeps both passes: the input one decides which parameter sets exist (SEI
and RPU arrays dropped there do propagate), the output one decides what the
sample entry asserts about them.

+++ Sources/PrismCore/DolbyVision/DolbyVisionRPUConverter.swift
@@ -100,7 +100,7 @@ final class DolbyVisionRPUConverter {
         #if canImport(Libdovi)
         return nal.withUnsafeBufferPointer { buffer -> [UInt8]? in
             guard let base = buffer.baseAddress,
-                  let rpu = dovi_parse_unspec62_nalu(base, UInt(buffer.count))
+                  let rpu = dovi_parse_unspec62_nalu(base, buffer.count)
             else { return nil }
             defer { dovi_rpu_free(rpu) }
